requirement.txt
aiosqlite==0.17.0
asyncpg==0.27.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
html5lib==1.1
idna==3.4
iso8601==1.1.0
pypika-tortoise==0.1.6
pytz==2023.3
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
tortoise-orm==0.19.3
typing_extensions==4.6.1
urllib3==2.0.2
webencodings==0.5.1
